requirements.txt
ibm-watson==4.7.1
python-decouple==3.3
